Start working on a worker process
git-svn-id: svn://acro.nym.se/projekt@310 2a83244e-8797-45b3-a2fb-03eebcce663c

+++ Mac/Almond/trunk/DraggableDetailView.m
@@ -46,7 +46,7 @@
         NSPoint tr = tl; tr.x += bounds.size.width - margin * 2;
         NSRect brect = NSMakeRect(bl.x, bl.y, tr.x - bl.x, tr.y - bl.y);
         
-        float lineDash[2];
+        CGFloat lineDash[2];
         lineDash[0] = 15.0;
         lineDash[1] = 9.0;
